Add writeSize metric
Summary:
Add writeSize variable to JMX so we can track average write sizes and see how
closely they correlate with put latency

Test Plan:
tested on cluster

DiffCamp Revision: 176576
Reviewed By: kannan
CC: kannan
Revert Plan:
OK

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181433 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/hadoop/hbase/io/hfile/HFile.java
@@ -170,13 +170,13 @@ public class HFile {
     DEFAULT_COMPRESSION_ALGORITHM.getName();
 
   // For measuring latency of "typical" reads and writes
-  private static volatile long readOps;
+  private static volatile int readOps;
   private static volatile long readTime;
-  private static volatile long writeOps;
+  private static volatile int writeOps;
   private static volatile long writeTime;
 
-  public static final long getReadOps() {
-    long ret = readOps;
+  public static final int getReadOps() {
+    int ret = readOps;
     readOps = 0;
     return ret;
   }
@@ -187,8 +187,8 @@ public class HFile {
     return ret;
   }
 
-  public static final long getWriteOps() {
-    long ret = writeOps;
+  public static final int getWriteOps() {
+    int ret = writeOps;
     writeOps = 0;
     return ret;
   }
